+++ assets/javascripts/kitten/components/navigation/header-nav/components/logged-out.js
@@ -1,8 +1,10 @@
-import React from 'react'
+import React, { useContext } from 'react'
 import { Context } from './context'
 
-export const LoggedOut = ({ children }) => (
-  <Context.Consumer>
-    {({ isLogged }) => (!isLogged ? children : null)}
-  </Context.Consumer>
-)
+export const LoggedOut = ({ children }) => {
+  const { isLogged } = useContext(Context)
+
+  if (isLogged) return null
+
+  return children
+}