+++ CHANGELOG.md
@@ -4,7 +4,22 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Fix: `Modal/Next`: Fix modal content height for decrepit versions of Safari.
+- Feature: `StickyContainer`: Add `"never"` value for `isSticky` prop.
+- Feature: `HeaderMenu`: Improve accessibility.
+- Feature: `HeaderNav.UserMenuNext`: Add `mobilePadding` prop.
+
+## [3.21.0] - 2021-06-21
+
+Feature:
+
+- `HeaderNav`: Add `BurgerMenuNext` and `UserMenuNext` subcomponents.
+
+Fixes:
+
+- `Modal/Next`: Fix modal content height for decrepit versions of Safari.
+- `VerticalCard`: Fix `description` PropType to node, allowing HTML parsing.
+- `CrowdfundingCard`: Fix `progress` PropTypes, allowing boolean values.
+- `Carousel`: Fix `Button` tiny size prop.
 
 ## [3.20.0] - 2021-06-17